requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23
tzdata==2022.4
urllib3==1.26.12
